backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pinecone
langchain
huggingface_hub
numpy
pydantic-settings
python-dotenv
supabase
openai
